capture message error catalog

+++ src/app/core/utilities/date.adapter.ts
@@ -25,7 +25,7 @@ export const MY_DATE_FORMATS = {
     // dateInput: { month: 'short', year: 'numeric', day: 'numeric' },
     dateInput: 'input',
     monthYearLabel: {year: 'numeric', month: 'short'},
-    dateA11yLabel: {year: 'numeric', month: 'long', day: 'numeric'},
-    monthYearA11yLabel: {year: 'numeric', month: 'long'},
+    dateA11yLabel: {year: 'numeric', month: 'short', day: 'numeric'},
+    monthYearA11yLabel: {year: 'numeric', month: 'short'},
   }
 };
